Update setup to use rst format (for PyPI)
PyPI docs don't like markdown, they use reStructuredText format.

The original setup.py had a workaround that involved converting the
file in real time, presumably so that people who don't know rst can
write the README in markdown, but this didn't make it to the release
on PyPI, so it's been released as raw markdown.

Github also understands rst, so it renders correctly there also.